Move AthenzIdentityVerifier to tls package
